requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-experimental
huggingface-hub
sentence-transformers
chromadb
pymysql
python-dotenv
mariadb